Disabling double initialization of the engine #1883
